fits: Defensive programming when there is no END header.
The keyword-search code assumes that it's dealing with a FITS file,
which must have an END header to terminate the header blocks. However,
if the FITS headers are in an NDF FITS airlock, the END header may be
absent (although it is frowned upon) as was the case for some UKIRT
Cassegrain-instrument data.  This modification also ends the search
loop when all the headers have been examined for the chosen keyword.